+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 
 ### Fixed
 - Negative zero liabilities in account summary
+- Dynamic account dropdown filter in transactions page
 
 ---